Create PNG file from plottable object

to_pdf plots an object directly into one ore multiple PDF files.

For a describr or describrGtable object, the object it first turned into a plottable object respecting the given size constraints (potential split-up) and then plotted to pmg using grid.draw().

For a list of describrGtable objects they are plotted directly to png files of suitable size.

to_png(x, name, ...)

# S3 method for describr
to_png(x, name, maxwidth = Inf, maxheight = Inf,
  dpi = 96, ...)

# S3 method for describrGtable
to_png(x, name, maxwidth = Inf, maxheight = Inf,
  dpi = 96, ...)

# S3 method for list
to_png(x, name, dpi = 96, ...)

Arguments

x
object to plot.

name
name used for output file(s), might be extended if multiple plots are generated and no ending is needed (e.g. 'my_file').

...
further options for specific methods.

maxwidth
maximal plotting width in inches (plain numeric).

maxheight
maximal plotting heigth in inches (plain numeric).

dpi
dots per inch